requirements.txt
flask
gunicorn
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-generativeai